requirements.txt
streamlit
pandas
camelot-py[cv]
pdfplumber
io